requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
duckduckgo-search 
beautifulsoup4
flask
flask-cors
